fix(recovery,a2r): structured-reviewer follow-ups for Findings 01 + 06
Two follow-ups surfaced by the post-fix structured-reviewer pass.

(1) Finding 06 missed a third site
----------------------------------
src/parsers/a2r/uft_a2r_parser.c:964 has the same uint8_t counts[256]
anti-pattern that F06 widened in src/recovery/. Today dormant because
the function's `count` parameter is uint8_t (max 255), so the lint did
not flag it (lint scans src/recovery/ only). Widened defensively to
match the F06 commits — same rationale, mechanical change.

(2) Finding 01: validity map now reachable by callers
-----------------------------------------------------
The reviewer correctly identified that the F01 fix allocated a
validity[] map internally to count authentic bytes, then freed it
without ever exposing it to the caller. Result: a sector that's
50 % sentinel-filled and a sector that's 100 % real-but-noisy both
emerged with confidence == 50, indistinguishable to downstream
forensic-report code. The doc-comment promised a contract the API
could not honour.

`uft_sector_recover_reconstruct` now takes an additional NULL-tolerant
out-parameter `uint8_t *out_validity`. Caller-allocated buffer
(sector_size bytes); on a successful recovery the function memcpys
the per-byte validity map into it before freeing the internal copy.
NULL keeps the old behaviour. Doc-comment updated to flag that
forensic tools should always pass a buffer.

No external callers of uft_sector_recover_reconstruct exist in src/
or include/ today (verified via grep), so the signature extension is
ABI-safe — nothing breaks because nothing depended on the old signature.

Verification
------------
gcc -fsyntax-only -Wall -Wextra clean on both files (mingw 13.1.0).
Pre-existing -Wtype-limits warnings in unrelated parse_rwcp_chunk /
parse_meta_chunk paths unaffected.
uft-recovery-integrity/lint_recovery.sh on uft_sector_recovery.c: clean.

Refs: structured-reviewer report after commit 45be506 (HTML logic
audit Findings 01 and 06 follow-up).

+++ src/parsers/a2r/uft_a2r_parser.c
@@ -959,25 +959,32 @@ a2r_error_t a2r_fuse_captures(const a2r_capture_t *captures,
     fused->data = malloc(ref->data_length);
     if (!fused->data) return A2R_ERR_ALLOC;
     
-    /* For each byte position, use majority vote */
+    /* For each byte position, use majority vote.
+     *
+     * counts[] and max_count widened from uint8_t to uint32_t (parallel
+     * to recovery Finding 06). `count` is uint8_t today so the bug is
+     * dormant, but the same future-proofing rationale applies: if the
+     * function signature is widened for a stress harness or external
+     * recovery wizard, a uint8_t counter wraps silently at 256 identical
+     * inputs and turns the majority winner into a near-loser. */
     for (uint32_t i = 0; i < ref->data_length; i++) {
-        uint8_t counts[256] = {0};
-        uint8_t max_val = 0;
-        uint8_t max_count = 0;
-        
+        uint32_t counts[256] = {0};
+        uint8_t  max_val   = 0;
+        uint32_t max_count = 0;
+
         for (uint8_t c = 0; c < count; c++) {
             if (i < captures[c].data_length) {
                 uint8_t val = captures[c].data[i];
                 counts[val]++;
                 if (counts[val] > max_count) {
                     max_count = counts[val];
-                    max_val = val;
+                    max_val   = val;
                 }
             }
         }
-        
+
         fused->data[i] = max_val;
-        
+
         /* Mark as weak if not unanimous */
         if (weak_mask && i / 8 < mask_size) {
             if (max_count < count) {
